remove sellerForm old component
